minor change for camlp5 8.03.00 compat
This breaks compat with previous versions of Camlp5

OCaml 5.2.0 supports raw identifiers, e.g. "\begin".  The Camlp5
pretty-printer needs to know which "identifier-like" strings are
keywords, so it can escape them as raw identifiers when
pretty-printing.  Since Camlp5 supports more than one syntax, that
means different syntaxes can have different sets of identifiers.
Hence, the lexer has to support exposing the set of keywords, and the
pretty-printer has to support being parameterized by the keywords.

matita creates a Camlp5 lexer, so it needs to be changed to have that
keyword-table slot (even though it seems it doesn't use it).